Moving shard positions to quickwit-ingest
