+++ quickwit/quickwit-indexing/src/models/mod.rs
@@ -30,7 +30,6 @@ mod processed_doc;
 mod publish_lock;
 mod publisher_message;
 mod raw_doc_batch;
-mod shard_positions;
 mod split_attrs;
 
 pub use indexed_split::{
@@ -50,8 +49,6 @@ pub use publish_lock::{NewPublishLock, PublishLock};
 pub use publisher_message::SplitsUpdate;
 use quickwit_proto::types::PublishToken;
 pub use raw_doc_batch::RawDocBatch;
-pub(crate) use shard_positions::LocalShardPositionsUpdate;
-pub use shard_positions::ShardPositionsService;
 pub use split_attrs::{create_split_metadata, SplitAttrs};
 
 #[derive(Debug)]
